Ändere Rückgabewert von Solve-Methoden in IDay-Implementierungen von int zu long; füge neue Day 07 Lösung hinzu

+++ Day_02/Part_02.cs
@@ -12,7 +12,7 @@ namespace Day_02
         const int MIN_DIFFERENCE = 1;
         const int MAX_DIFFERENCE = 3;
 
-        public int Solve(string[] pInput)
+        public long Solve(string[] pInput)
         {
             var count = 0;
             foreach (var line in pInput)
